对未达到的指标进行代码补充和修改
1. P1 权限审批链路不完整
   - 问题：高优任务缺少审批门禁；任务详情/取消存在越权风险。
   - 已修复：
     - 高优任务门禁（>= threshold 需审批角色）
     - 非审批角色活动任务配额限制
     - 任务详情/取消加入“任务归属校验（管理员例外）”
2. P1 分布式锁释放非原子
   - 问题：get+delete 释放锁有竞态。
   - 已修复：Lua 脚本原子 compare-and-del。
3. P2 BestFit 未显式按“显存浪费率”决策
   - 问题：此前主要按最小显存排序，未显式计算浪费率。
   - 已修复：按 waste=(gpuMem-required)/gpuMem 最小化并记录 wasteRate。
4. P2 大屏遥测 KPI 不全
   - 问题：缺少“各优先级队列长度”“历史调度耗时”。
   - 已修复：新增并通过 WebSocket 推送：
     - queueLengthByPriority
     - avgDispatchLatencySeconds
     - avgTurnaroundSeconds
     - 每卡 usedMemoryGbByGpu / remainingMemoryGbByGpu
5. P2 RBAC 表达式存在前缀误用风险
   - 问题：hasRole('ROLE_ADMIN') 可能导致 ROLE_ROLE_*。
   - 已修复：改为 hasAuthority/hasAnyAuthority。

+++ src/test/java/com/sakurain/gpuscheduler/scheduler/TaskDispatcherTest.java
@@ -72,7 +72,7 @@ class TaskDispatcherTest {
 
     @Test
     void testDispatch_SuccessfulAllocation() {
-        when(priorityQueue.dequeue()).thenReturn(100L, null);
+        when(priorityQueue.dequeue()).thenReturn(100L, (Long) null);
         when(taskMapper.selectById(100L)).thenReturn(queuedTask);
         when(gpuAllocator.allocate(queuedTask)).thenReturn(Optional.of(idleGpu));
 
@@ -106,7 +106,7 @@ class TaskDispatcherTest {
 
     @Test
     void testDispatch_TaskNotFound() {
-        when(priorityQueue.dequeue()).thenReturn(100L, null);
+        when(priorityQueue.dequeue()).thenReturn(100L, (Long) null);
         when(taskMapper.selectById(100L)).thenReturn(null);
 
         dispatcher.dispatchOnce();
@@ -122,7 +122,7 @@ class TaskDispatcherTest {
                 .status(TaskStatus.RUNNING.getCode())
                 .build();
 
-        when(priorityQueue.dequeue()).thenReturn(100L, null);
+        when(priorityQueue.dequeue()).thenReturn(100L, (Long) null);
         when(taskMapper.selectById(100L)).thenReturn(runningTask);
 
         dispatcher.dispatchOnce();
@@ -187,7 +187,7 @@ class TaskDispatcherTest {
     }
     @Test
     void testDispatch_AllocationFailure() {
-        when(priorityQueue.dequeue()).thenReturn(100L, null);
+        when(priorityQueue.dequeue()).thenReturn(100L, (Long) null);
         when(taskMapper.selectById(100L)).thenReturn(queuedTask);
         when(gpuAllocator.allocate(queuedTask)).thenReturn(Optional.of(idleGpu));
         doThrow(new RuntimeException("Database error"))
@@ -200,7 +200,7 @@ class TaskDispatcherTest {
 
     @Test
     void testDispatch_EstimatedTimeCalculation() {
-        when(priorityQueue.dequeue()).thenReturn(100L, null);
+        when(priorityQueue.dequeue()).thenReturn(100L, (Long) null);
         when(taskMapper.selectById(100L)).thenReturn(queuedTask);
         when(gpuAllocator.allocate(queuedTask)).thenReturn(Optional.of(idleGpu));
